refactor(test): Move tokenizer tests to `test/test_tokenizer/`
